Experiment with refactoring index.js files
Added to the nav.js, a function that should perform the same steps as the many index.js boilerplate files that exist.
